requirements.txt
Django>=4.2,<5.1
Pillow>=10.0
psycopg[binary]>=3.1   # or psycopg2-binary>=2.9
djangorestframework
djangorestframework-simplejwt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0  
django-cors-headers

